Arreglo % de similitud postulaciones

+++ server/routes/empleado.py
@@ -644,7 +644,7 @@ def postularme(id_oferta):
         is_apto=aptitud_cv,
         fecha_postulacion=datetime.now(timezone.utc),
         estado_postulacion="pendiente",
-        porcentaje_similitud=porcentaje * 100,
+        porcentaje_similitud=int(porcentaje*100),
     )
 
     db.session.add(nueva_postulacion)
